requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
CacheControl==0.12.11
cachelib==0.10.2
cachetools==5.3.0
camelot-py==0.9.0
canvas==0.3
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
convertdate==2.4.0
cryptography==40.0.2
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.6
distro==1.8.0
Django==4.2.1
django-allauth==0.54.0
django-debug-toolbar==4.1.0
django-pdfkit==0.3.1
docx==0.2.4
dominate==2.7.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executor==23.2
fasteners==0.18
filelock==3.12.0
firebase-admin==6.1.0
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-WTF==1.1.1
fonttools==4.39.4
freetype-py==2.3.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-cloud-vision==3.4.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
hijri-converter==2.3.1
holidays==0.23
html-to-json==2.0.0
html5==0.0.9
httplib2==0.22.0
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
msgpack==1.0.5
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pdf2image==1.16.3
pdfkit==0.6.0
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pipenv==2023.5.19
platformdirs==3.5.1
poppler-utils==0.1.0
property-manager==3.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.23.0
pycparser==2.21
PyJWT==2.6.0
PyMeeus==0.5.12
PyMuPDF==1.22.3
pyparsing==3.0.9
PyPDF2==3.0.1
PyPDF4==1.27.0
pyreadline3==3.4.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
render==1.0.0
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
rlPyCairo==0.2.0
rsa==4.9
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tabula-py==2.7.0
tesseract==0.1.3
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
utils==1.0.1
verboselogs==1.7
virtualenv==20.23.0
virtualenv-clone==0.5.7
visitor==0.1.3
Wand==0.6.11
webencodings==0.5.1
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.0.1
wtforms-html5==0.6.1
